fix: cleanup + feed connection cta
