+++ .env.development
@@ -2,4 +2,3 @@ NEXT_PUBLIC_ARWEAVE_CDN_HOST=/api/arweave/txt
 NEXT_PUBLIC_MARKETPLACE_ENABLED=true
 FEATURED_STOREFRONTS_URL="https://nftstorage.link/ipfs/bafkreifay7aebzkasogdrc2l75tvln45qc5noygi6w5bjkzja2giqw5nzq"
 NEXT_PUBLIC_ENVIRONMENT=development
-# NEXT_PUBLIC_INDEXER_GRAPHQL_URL="https://graph.holaplex.com/v1"
