+++ pages/nfts/[address].tsx
@@ -1,4 +1,3 @@
-/* eslint-disable react-hooks/rules-of-hooks */
 import React, { useEffect, useState } from 'react';
 import { GetServerSideProps } from 'next';
 import { ListingReceipt, useNftMarketplaceLazyQuery } from '../../src/graphql/indexerTypes';
